+++ src/CMakeLists.txt
@@ -31,7 +31,7 @@ function(compile_lib LIB_NAME DIR DEPEND_FILE_LIST)
     set(FILE_STR_LIST "")
     file(GLOB_RECURSE SRC_FILES ${DIR}/*.cc)
     foreach(SRC_FILE ${SRC_FILES})
-        if (NOT SRC_FILE MATCHES ".*_test.cc")
+        if (NOT SRC_FILE MATCHES ".*_test.cc" AND NOT SRC_FILE MATCHES ".*_bm.cc")
             set(FILE_STR_LIST "${FILE_STR_LIST} ${SRC_FILE}")
         endif()
     endforeach()
